Always wait for the reply from the OpenURI portal before quitting
The flow of events after clicking on the icon is something like this:

The sandboxed application sends D-Bus message to the portal's backend,
which searches for suitable applications to handle with the URI request.
Once found, the backend sends another D-Bus message to the frontend (a
different process) and the apps selector can be shown.

Now, once the user selects what to do (being "Cancel" an option), a
reply message is sent back to the backend and then relayed to the
originating application, inside the sandbox.

For this reason, it's not a particularly good decision to simply fire
up the D-Bus message to ask for the URI to be opened and then exit
right away. Instead, we should wait for the reply coming from the
portal and only exit once that's received.

Also, I believe this was also the reason why sometimes clicking on the
Dropbox icon would do nothing, as there's a chance that the portal
won't properly handle the message is we exit too soon.

https://phabricator.endlessm.com/T15032