Own 'com.dropbox.Client' on the session bus for the first instance
It is not possible to check from a sandbox whether there's another
'dropbox' process running in the system since we can't check PIDs
outside of the sandbox, and even if we checked the dropbox.pid file
that would refer to a PID inside a different sandbox.

For this reason, we better own a name on the session bus, which is
something we can easily check from any other instance of this flatpak
application, and use the 'name_acquired' and 'name_lost' to decide
whether we only want to open the Dropbox directory or also to spawn
a new running instance of the 'dropbox' daemon.

https://phabricator.endlessm.com/T15115