Always launch the daemon if the Dropbox folder is not found
When the Dropbox folder is not found we need to launch the daemon, even
if it is already running. The reason is that Dropbox has changed and now
it opens the browser for the user authentication, instead of its own
application. So, if the user just closes the browser, or an error
occurred and it couldn't be opened, we're left in a state where the
daemon is running but Dropbox has not been configured. Thus, if we do
not run the daemon again, the user will not be able to log in again.

https://phabricator.endlessm.com/T16226