Set monitors to better manage when to open the Dropbox dir
The configured Dropbox directory may not be immediately available when
the daemon is started, thus readily opening at that moment may not be
very effective.

This patch sets up a monitor for the configuration file and, once the
directory is configured, starts watching it, opening it only when it is
effectively created.

https://phabricator.endlessm.com/T16271