Replace a running instance of the app when no configuation is set
As set in b69ef501b356965287285faf6912ce8f1901c0dd, we now launch a new
Dropbox daemon when there is no configuration available, even if there
is already an instance of the app running.

However, we ideally want to have just one instance of Dropbox running,
so this patch changes how the DBus name is owned, effectively replacing
the owner under the conditions mentioned above: no configuration found
and a running instance already.
This works because now the app also quits if it loses the ownership of
the DBus name.

https://phabricator.endlessm.com/T16271