Terminate the Dropbox daemon when quitting the app
Otherwise the daemon keeps running in the background, and after starting
more instances we're left with multiple daemons running, which leads to
memory consumption and other potential problems.

This problem didn't happen when aborting the app (Ctrl+C) but when it
quit by itself because a new instance took its DBus ownership: to
reproduce launch Dropbox, remove the configuration folder (~/.dropbox)
and launch Dropbox again.

https://phabricator.endlessm.com/T16271