Use the new OpenFile method from the OpenURI portal to open the Dropbox's directory
In the 0.6 version of the portal, the OpenURI method has been changed not to handle
file:// URIs, which now need to be handled via the new OpenFile method, passing a
file descriptor to it instead of the path as a "proof" of the app being able to
reach such path from inside the sandbox.

This commit, changes the call in our wrapper that makes it possible to open the
Dropbox's directory when starting the app (e.g. by clicking on the icon), so that
the new method is used, and a file descriptor is passed.

https://phabricator.endlessm.com/T16415