Add fallback code to use the OpenURI method if OpenFile is not available
If the user updates the Dropbox application before updating the OS, the
current logic won't work because it will try to use an OpenFile method
that doesn't exist yet in the older (< 0.6) version of the portal.

To deal with this situation, we add some fallback code so that the old
OpenURI method is used if an UnknownMethod exception is raised while
trying to call OpenFile, in which case is reasonable to assume that
it happens because of the older version of the portals being available.

https://phabricator.endlessm.com/T16415