Make sure to monitor the actual (binary) daemon that dropbox will run
Previously, we were not actually exitting from the wrapper script when
dropbox's daemon has been successfully run once, regardless of whether
such daemon has been uncleanly terminated/killed or cleanly exited from
the icon in the status bar area, since nothing would quit the mainloop
in that case. Because of this, combined with the fact that this app
handles uniqueness by checking whether the bus name is owned, we could
easily end up in a situation where the dropbox's daemon would never run
in a session, if the app have been previously run at least once and the
daemon terminated (cleanly or uncleanly) ever since.

To fix this problem, we now distinguish between the dropbox's actual
"daemon" (a binary that gets executed by means of calling a launcher
script) and the "launcher" (a shell script whose only responsibility
is to provide a clear entry point to launch the daemon). By doing that,
we can make sure that we wait until the launcher's process gets replaced
with the one for the actual daemon before continuing, so that we can
install a simple monitoring system that will ensure the app quits as
soon as the dropbox's daemon is terminated, releasing the bus name.

https://phabricator.endlessm.com/T21400