Handle the case where the initial launcher does not get terminated
In previous commits, we wrote code assuming what seemed to be the case
as per the dropbox's daemon behaviour: that the launcher will start
the daemon and then eventually exit itself, leaving the daemon alone.

However, it turns out now that Dropbox does not always do that, and
that sometimes it keeps the initial launcher alive forever (or for
quite a long time at least), meaning that we can't simply wait() for
that initial process to finish unless we want to get artificially
blocked at that point (e.g. no Dropbox directory showing up).

To fix this, let's change the logic in the secondary thread to handle
also the launcher's lifetime, consider the need to monitor that other
binary daemon directly as something "optional" (only if the process
is alive when the launcher exits) and make sure that the application
exits whenever both those two potentially running procesess exit.

https://phabricator.endlessm.com/T21400