Disable auto-updates in the ~/.dropbox-dist directory
This is all a bit obscure and still a bit of "guessing territory", but
it looks like some of the inconsistent issues we have been observing
could have been caused by the fact that Dropbox sometimes creates a
copy of the .dropbox-dist directory installed with the flatpak app
under the user's home directory.

It's still a bit unclear how this works, but it seems like this could
be some kind of auto-update mechanism and that, at least under some
situations, Dropbox might decide to run the actual daemon off that
directory instead of the one installed with the flatpak, which can
lead to bizarre situations (specially if versions don't match).

As a tentative fix to deal with this type of situation, let's attempt
to "disable" the usage of this directory by making sure it's empty
and by making it non writable.

https://phabricator.endlessm.com/T21400