Override xdg-open from the runtime with our very own version
The version of xdg-open from the runtime does not convert absolute
paths into file:// URIs, and because of that subsequent calls (via
gvfs-open, called from xdg-open) to the OpenURI portal won't work,
as that only accepts URIs.

Since this app is build against the legacy com.endlessm.Platform
runtime, and that one includes a older version of GLib (2.50.2),
it probably makes sense to ship our own version of xdg-open to
make sure it does the right thing, making the most that we already
have the code in place to deal with either the OpenURI or the
OpenFile methods (depending on the version of the portal installed).

This commit, provides a small python script that hooks onto the
same module already used by eos-dropbox-app to communicate with
the methods of the OpenURI portal.

https://phabricator.endlessm.com/T15069