portallauncher: Ensure to always pass a valid URI to the OpenURI method
Without this patch, falling back to the OpenURI() method when the
OpenFile() one was not available to open a local path will fail,
since the file:// prefix needs to be explicitly added.

Also, refactor a bit the way we create the private attributes of the
PortalLauncher, to avoid duplicating too much code unnecessarily.

https://phabricator.endlessm.com/T15069