Remove our version of xdg-open, no longer needed
This was added to deal with situations where the user might have
installed either an older version (pre 0.6) of the portals or
the new one, to gracefully deal with OpenFile or OpenURI methods,
as well as to make sure the OpenURI portal was being used when
xdg-open was invoked from the Dropbox daemon.

Since we're using now a newer runtime which ships a flatpak-specific
version of xdg-open that interacts correctly with the portal, and
the minimum version of the portals users in 3.4+ versions of EOS
will be good enough, there's no need to keep this around so drop it.

https://phabricator.endlessm.com/T22890