eos-dropbox-app: Fix monitoring of the dropbox daemon
This was previously implemented by waiting for the launcher to die
and then finding the daemon's process to monitor it, but this does
not work anymore since now the original launcher does not really die,
but gets replaced with the daemon's process... meaning that the
monitoring is never setup until the original process exits.

So, simplify everything a bit by simply relying in the blocking call
to Popen and then .wait() afterwards to know when to kill the flatpak
process after the daemon dies, and get rid of so many weird things we
added in the past to deal with the previous situation.

https://phabricator.endlessm.com/T22823