eos-dropbox-app: Check for additional daemon process before quitting
This is fairly weird, but we found out that in some situations an
additional process for the dropbox daemon gets spawned AFTER the
original launcher process has quit, which breaks the way uniqueness
is managed since the wrapper will quit and release the bus name.

As a quick attempt to try to handle this situation, let's add an
extra check after the call to wait() finishes to look for instances
of the dropbox daemon running, and wait for it in that case.

https://phabricator.endlessm.com/T22823